requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain>=0.1.0
langchain-community>=0.0.16
huggingface-hub>=0.23.0
sentence-transformers>=2.6.0
transformers>=4.39.0
langchain-huggingface==0.0.1
accelerate==0.24.1
llama-cpp-python==0.2.19
pypdf==3.17.1
docx2txt==0.8
python-pptx==0.6.21
faiss-cpu==1.7.4
torch>=2.1.1
torchvision>=0.16.0
tqdm==4.66.1
pydantic==2.5.2
python-dotenv>=1.0.0